Moved data collection code to data module.
